moving pngToSvg logic to download util
